refactor: update abi in test and move var declaration
